backend/api/flask/requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
requests==2.32.3
python-dotenv==1.0.1